nftables-api/requirements.txt
autopep8==1.6.0
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
python-dotenv==0.20.0

